Note:
This script will allow you to copy multiple files within many folders and paste them into a specific folder. It is best used for copying tons of content in multiple folders instead of doing it 1 by 1. For best case use, copy all the folders that contain the content you need to copy into a folder, and run this script to do its work.